:goal_net: :bulb: Improve error handling
"backup" and "synchronize_backup" can fail to write to stdout.
You can test this by redirecting the output to "/dev/full".

In the code of "ninja_bootstrap", you can now see a use case for IIFE in Rust.
Most often, an IIFE can be replaced with a block expression.
But, when the question mark operator is used, the behavior is different.
(IIFE means Immediately Invoked Function Expression.)